+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/Views/ActionMapsView.cs
@@ -30,22 +30,16 @@ namespace UnityEngine.InputSystem.Editor
                 treeViewItem.label.text = mapData.mapName;
                 treeViewItem.EditTextFinishedCallback = newName => ChangeActionMapName(i, newName);
                 treeViewItem.EditTextFinished += treeViewItem.EditTextFinishedCallback;
-                treeViewItem.DeleteCallback = _ => DeleteActionMap(i);
-                treeViewItem.DuplicateCallback = _ => DuplicateActionMap(i);
-                treeViewItem.OnDeleteItem += treeViewItem.DeleteCallback;
-                treeViewItem.OnDuplicateItem += treeViewItem.DuplicateCallback;
                 treeViewItem.userData = i;
                 element.SetEnabled(!mapData.isDisabled);
 
-                ContextMenu.GetContextMenuForActionMapItem(this, treeViewItem);
+                ContextMenu.GetContextMenuForActionMapItem(this, treeViewItem, i);
             };
             m_ListView.makeItem = () => new InputActionMapsTreeViewItem();
             m_ListView.unbindItem = (element, i) =>
             {
                 var treeViewElement = (InputActionMapsTreeViewItem)element;
                 treeViewElement.Reset();
-                treeViewElement.OnDeleteItem -= treeViewElement.DeleteCallback;
-                treeViewElement.OnDuplicateItem -= treeViewElement.DuplicateCallback;
                 treeViewElement.EditTextFinished -= treeViewElement.EditTextFinishedCallback;
             };
 
@@ -58,10 +52,14 @@ namespace UnityEngine.InputSystem.Editor
             m_ListView.RegisterCallback<ExecuteCommandEvent>(OnExecuteCommand);
             m_ListView.RegisterCallback<ValidateCommandEvent>(OnValidateCommand);
             m_ListView.RegisterCallback<PointerDownEvent>(OnPointerDown, TrickleDown.TrickleDown);
+
+            // ISXB-748 - Scrolling the view causes a visual glitch with the rename TextField. As a work-around we
+            // need to cancel the rename operation in this scenario.
+            m_ListView.RegisterCallback<WheelEvent>(e => InputActionMapsTreeViewItem.CancelRename(), TrickleDown.TrickleDown);
+
             var treeView = root.Q<TreeView>("actions-tree-view");
             m_ListView.AddManipulator(new DropManipulator(OnDroppedHandler, treeView));
             m_ListView.itemIndexChanged += OnReorder;
-
 
             CreateSelector(Selectors.GetActionMapNames, Selectors.GetSelectedActionMap, (actionMapNames, actionMap, state) => new ViewState(actionMap, actionMapNames, state.GetDisabledActionMaps(actionMapNames.ToList())));
 
@@ -108,15 +106,23 @@ namespace UnityEngine.InputSystem.Editor
             if (element == null)
                 return;
             ((InputActionMapsTreeViewItem)element).FocusOnRenameTextField();
-            m_EnterRenamingMode = false;
         }
 
-        private void DeleteActionMap(int index)
+        internal void RenameActionMap(int index)
+        {
+            m_ListView.ScrollToItem(index);
+            var element = m_ListView.GetRootElementForIndex(index);
+            if (element == null)
+                return;
+            ((InputActionMapsTreeViewItem)element).FocusOnRenameTextField();
+        }
+
+        internal void DeleteActionMap(int index)
         {
             Dispatch(Commands.DeleteActionMap(index));
         }
 
-        private void DuplicateActionMap(int index)
+        internal void DuplicateActionMap(int index)
         {
             Dispatch(Commands.DuplicateActionMap(index));
         }
@@ -138,6 +144,7 @@ namespace UnityEngine.InputSystem.Editor
 
         private void ChangeActionMapName(int index, string newName)
         {
+            m_EnterRenamingMode = false;
             Dispatch(Commands.ChangeActionMapName(index, newName));
         }
 
@@ -163,10 +170,10 @@ namespace UnityEngine.InputSystem.Editor
                         break;
                     case CmdEvents.Delete:
                     case CmdEvents.SoftDelete:
-                        ((InputActionMapsTreeViewItem)selectedItem).DeleteItem();
+                        DeleteActionMap(m_ListView.selectedIndex);
                         break;
                     case CmdEvents.Duplicate:
-                        ((InputActionMapsTreeViewItem)selectedItem).DuplicateItem();
+                        DuplicateActionMap(m_ListView.selectedIndex);
                         break;
                     case CmdEvents.Copy:
                         CopyItems();