+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/Views/ControlSchemesView.cs
@@ -25,7 +25,7 @@ namespace UnityEngine.InputSystem.Editor
                 InputActionsEditorConstants.ControlSchemeEditorViewUxml);
 
             var controlSchemeVisualElement = controlSchemeEditor.CloneTree();
-            controlSchemeVisualElement.Q<Button>(kCancelButton).clicked += Close;
+            controlSchemeVisualElement.Q<Button>(kCancelButton).clicked += Cancel;
             controlSchemeVisualElement.Q<Button>(kSaveButton).clicked += SaveAndClose;
             controlSchemeVisualElement.Q<TextField>(kControlSchemeNameTextField).RegisterCallback<FocusOutEvent>(evt =>
             {
@@ -50,7 +50,7 @@ namespace UnityEngine.InputSystem.Editor
             m_ModalWindow.Add(popupWindow);
             root.Add(m_ModalWindow);
             m_ModalWindow.StretchToParentSize();
-            m_ModalWindow.RegisterCallback<ClickEvent>(evt => Close());
+            m_ModalWindow.RegisterCallback<ClickEvent>(evt => CloseView());
             popupWindow.RegisterCallback<ClickEvent>(evt => evt.StopPropagation());
 
             m_ListView = controlSchemeVisualElement.Q<MultiColumnListView>(kControlSchemesListView);
@@ -111,12 +111,28 @@ namespace UnityEngine.InputSystem.Editor
 
         private void SaveAndClose()
         {
+            // Persist the current ControlScheme values to the SerializedProperty
             Dispatch(ControlSchemeCommands.SaveControlScheme(m_NewName, m_UpdateExisting));
-            Close();
+            CloseView();
         }
 
-        private void Close()
+        private void Cancel()
         {
+            // Reload the selected ControlScheme values from the SerilaizedProperty and throw away any changes
+            Dispatch(ControlSchemeCommands.ResetSelectedControlScheme());
+            CloseView();
+        }
+
+        private void CloseView()
+        {
+            // Closing the View without explicitly selecting "Save" or "Cancel" holds the values in the
+            // current UI state but won't persist them; the Asset Editor state isn't dirtied.
+            //
+            // This means accidentally clicking outside of the View (closes the dialog) won't immediately
+            // cause loss of work: the "Edit Control Scheme" option can be selected to re-open the View
+            // the changes retained. However, if a different ControlScheme is selected or the Asset
+            // Editor window is closed, then the changes are lost.
+
             m_NewName = "";
             OnClosing?.Invoke(this);
         }